requirements.txt
gender-guesser==0.4.0
python-twitter==3.3
